Refactor operations.ts
Split monolithic file into three files so it's easier to reason about.
Added barrel files to clarify external APIs used by the emitter.
Moved shared functions to helpers.ts.
Simplified a few bits of code to avoid more shared helpers.
