+++ .chronus/changes/go-op-refactor-2026-7-6-15-44-32.md
@@ -1,0 +1,7 @@
+---
+changeKind: internal
+packages:
+  - "@azure-tools/typespec-go"
+---
+
+Split operations.ts into three files to better contain dependent code.
