+++ packages/typespec-go/src/codegen/core/helpers.ts
@@ -121,7 +121,7 @@ export function formatParameterTypeName(
 }
 
 // sorts parameters by their required state, ordering required before optional
-export function sortParametersByRequired(
+function sortParametersByRequired(
   a: go.ClientParameter | go.ParameterGroup,
   b: go.ClientParameter | go.ParameterGroup,
 ): number {
@@ -306,6 +306,28 @@ export function getParamName(param: go.MethodParameter): string {
   return paramName;
 }
 
+export function fixUpMethodName(method: go.MethodType): string {
+  switch (method.kind) {
+    case "lroMethod":
+    case "lroPageableMethod":
+      return `Begin${method.name}`;
+    case "pageableMethod": {
+      let N = "N";
+      let name = method.name;
+      if (method.name[0] !== method.name[0].toUpperCase()) {
+        // the method isn't exported; don't export the pager ctor
+        N = "n";
+        // ensure correct casing of the emitted function name e.g.,
+        // "listThings" -> "newListThingsPager"
+        name = name[0].toUpperCase() + name.substring(1);
+      }
+      return `${N}ew${name}Pager`;
+    }
+    case "method":
+      return method.name;
+  }
+}
+
 // converts the Go code model encoding type to the type name in the standard library
 export function formatBytesEncoding(enc: go.BytesEncoding): string {
   if (enc === "URL") {
@@ -379,6 +401,16 @@ export function formatParamValue(
   return formatValue(paramName, param.type, imports);
 }
 
+/**
+ * returns the receiver definition for a client
+ *
+ * @param receiver the receiver for which to emit the definition
+ * @returns the receiver definition
+ */
+export function getClientReceiverDefinition(receiver: go.Receiver<go.Client>): string {
+  return `(${receiver.name} ${receiver.byValue ? "" : "*"}${receiver.type.name})`;
+}
+
 export function getDelimiterForCollectionFormat(cf: go.CollectionFormat): string {
   switch (cf) {
     case "csv":
@@ -392,6 +424,28 @@ export function getDelimiterForCollectionFormat(cf: go.CollectionFormat): string
     default:
       throw new CodegenError("InternalError", `unhandled CollectionFormat ${cf}`);
   }
+}
+
+export function getMediaFormat(type: go.WireType, mediaType: "JSON" | "XML", param: string): string {
+  let marshaller: "JSON" | "XML" | "ByteArray" = mediaType;
+  let format = "";
+  if (type.kind === "encodedBytes") {
+    marshaller = "ByteArray";
+    format = `, runtime.Base64${type.encoding}Format`;
+  }
+  return `${marshaller}(${param}${format})`;
+}
+
+export function isMapOfDateTime(
+  paramType: go.WireType,
+): { format: go.TimeFormat; utc: boolean } | undefined {
+  if (paramType.kind !== "map") {
+    return undefined;
+  }
+  if (paramType.valueType.kind !== "time") {
+    return undefined;
+  }
+  return { format: paramType.valueType.format, utc: paramType.valueType.utc };
 }
 
 export function formatValue(
@@ -502,19 +556,6 @@ export function formatLiteralValue(value: go.Literal, withCast: boolean): string
       return `"${value.literal}"`;
     case "time":
       return `"${value.literal}"`;
-  }
-}
-
-// returns true if at least one of the responses has a schema
-export function hasSchemaResponse(method: go.MethodType): boolean {
-  switch (method.returns.result?.kind) {
-    case "anyResult":
-    case "modelResult":
-    case "monomorphicResult":
-    case "polymorphicResult":
-      return true;
-    default:
-      return false;
   }
 }
 
@@ -1418,7 +1459,7 @@ export function camelCase(identifier: string | Array<string>): string {
   return `${naming.uncapitalize(identifier[0])}${pascalCase(identifier.slice(1))}`;
 }
 
-export function pascalCase(identifier: string | Array<string>): string {
+function pascalCase(identifier: string | Array<string>): string {
   return identifier === undefined
     ? ""
     : typeof identifier === "string"
