+++ packages/typespec-go/src/codegen/core/index.ts
@@ -1,0 +1,21 @@
+/*---------------------------------------------------------------------------------------------
+ *  Copyright (c) Microsoft Corporation. All rights reserved.
+ *  Licensed under the MIT License. See License.txt in the project root for license information.
+ *--------------------------------------------------------------------------------------------*/
+
+export { generateClientFactory } from "./client-factory.js";
+export { generateCloudConfig } from "./cloud-config.js";
+export { generateConstants } from "./constants.js";
+export { generateExamples } from "./example.js";
+export { generateGoModFile } from "./gomod.js";
+export { setCustomHeaderText } from "./helpers.js";
+export { generateInterfaces } from "./interfaces.js";
+export { generateLicenseTxt } from "./license.js";
+export { generateMetadataFile } from "./metadata.js";
+export { generateModels } from "./models.js";
+export { generateOperations } from "./operations.js";
+export { generateOptions } from "./options.js";
+export { generatePolymorphicHelpers } from "./polymorphics.js";
+export { generateResponses } from "./responses.js";
+export { generateVersionInfo } from "./version.js";
+export { generateXMLAdditionalPropsHelpers } from "./xml-additional-props.js";
