+++ packages/typespec-go/src/codegen/fake/index.ts
@@ -1,0 +1,7 @@
+/*---------------------------------------------------------------------------------------------
+ *  Copyright (c) Microsoft Corporation. All rights reserved.
+ *  Licensed under the MIT License. See License.txt in the project root for license information.
+ *--------------------------------------------------------------------------------------------*/
+
+export { generateServerFactory } from "./factory.js";
+export { generateServers } from "./servers.js";
